HGMSMatcher class implementation
HGMSMatcher class. it is essentially the same as GMS with hierarchical logic added on top. Two main functions, getInlierMask and run, were changed significantly, and the variable that holds the original gridsize was changed to a vector list that stores all the hierarchical grid sizes needed to go from coarse to fine grid. Almost all the other functions are the same, with very minor adjustments to accommodate the main changes.